refactor: move Form export to index.ts; update schema validation; adjust imports
