+++ src/pages/login/components/form/styles.module.css
@@ -1,0 +1,28 @@
+.form {
+    display: grid;
+    gap: 24px;
+    padding: 30px 0 0 42px;
+    justify-items: flex-start;
+}
+
+.field {
+    min-width: 364px;
+}
+
+.error {
+    color: var(--colors-interface-orange);
+}
+
+@media all and (max-width: 415px) {
+    .form {
+        padding: 40px 0 0 10px;
+    }
+
+    .button {
+        margin-top: 12px;
+    }
+
+    .field {
+        min-width: 98%;
+    }
+}
