+++ src/pages/login/styles.module.css
@@ -4,26 +4,11 @@
 
 .title {
   margin: 0;
-  padding: 40px 18px 12px;
+  padding: 40px 30px 12px;
   font-family: 'Golos', sans-serif;
   font-size: 24px;
   line-height: 29px;
   color: var(--colors-interface-primary);
-}
-
-.form {
-  display: grid;
-  gap: 24px;
-  padding: 30px 0 0 18px;
-  justify-items: flex-start;
-}
-
-.field {
-  min-width: 364px;
-}
-
-.error {
-  color: var(--colors-interface-orange);
 }
 
 @media all and (max-width: 415px) {
@@ -34,17 +19,5 @@
   .title {
     padding: 0 10px 20px;
     font-size: 16px;
-  }
-
-  .form {
-    padding: 40px 0 0 10px;
-  }
-
-  .button {
-    margin-top: 12px;
-  }
-
-  .field {
-    min-width: 98%;
   }
 }
